2.0a.rev26
Functionalized the "Confirm" button of the modal diaglogs. Trash and
Config are hooked up, share does nothing right now except hide the
modal.

Editing and adding with "newly created group" now working.

Need form validation on the forms though. Ok should not go through if
form is invalid.

Also found that if focus is in a select box, Esc does not work to hide
the modal, need to fix that.