2.0a.rev29
I was getting that `invalid in` on cutils line 23 in the low level hook,
i just set it to Object.proto....call method and its fixed up

As my method in sync worker turned out to be async, i moved it to async.
I then fixed up sync worker so low level hook works there. It seems
GetMessage while it blocks is dispatching the "sent messages" and its
never returning because it is not getting a "posted message" as per the
first summary block on msdn docs of GetMessage. So I use PostMessage in
the callback to make that GetMessage return. But GetMessage blocking is
what makes the low level hook work! aH!

Some styling on modal when create